Time: 85 ms (72.47%) | Memory: 44.5 MB (89.68%) - LeetSync

+++ 2206-detonate-the-maximum-bombs/detonate-the-maximum-bombs.java
@@ -24,6 +24,6 @@ class Solution {
         long dx = cell1[0]-cell2[0];
         long dy = cell1[1]-cell2[1];
         long range = cell1[2];
-        return (long) dx*dx + dy*dy <= (long) range*range;
+        return dx*dx + dy*dy <= range*range;
     }
 }
